Improved testing links in docs.
