+++ tools/gulptasks/test.js
@@ -182,7 +182,6 @@ function checkDemosConsistency() {
  */
 function checkDocsConsistency() {
     const FS = require('fs');
-    const glob = require('glob');
     const LogLib = require('./lib/log');
 
     const sidebar = require('../../docs/sidebars.js');
@@ -223,29 +222,6 @@ function checkDocsConsistency() {
     }
 
     // Check links and references to samples
-    glob.sync(process.cwd() + '/docs/**/*.md').forEach(file => {
-        const md = FS.readFileSync(file),
-            regex = /(https:\/\/jsfiddle.net\/gh\/get\/library\/pure\/highcharts\/highcharts\/tree\/master\/samples|https:\/\/www.highcharts.com\/samples\/embed)\/([a-z0-9\-]+\/[a-z0-9\-]+\/[a-z0-9\-]+)/gu;
-
-        const error404s = [];
-
-        let match;
-        while ((match = regex.exec(md))) {
-            const sample = match[2].replace(/\/$/u, '');
-            try {
-                FS.statSync(`samples/${sample}/demo.js`);
-            } catch (error) {
-                error404s.push({ file, sample });
-            }
-        }
-
-        if (error404s.length) {
-            throw new Error(
-                'Rotten links\n' + JSON.stringify(error404s, null, '  ')
-            );
-        }
-
-    });
 
 }
 
@@ -491,4 +467,4 @@ Available arguments for 'gulp test':
     });
 }
 
-gulp.task('test', gulp.series('scripts', test));
+gulp.task('test', gulp.series('test-links-docs', 'scripts', test));
